fixed word and removed super class call
calling super in onRequestPermissionsResult should be avoided in the overridden method now unless the plugin author wants an easy way to return the plugin states to the saved call. I think we do want to encourage that, but this example doesn't currently show how to detect the original saved call type and respond if it was just a `checkPermissions` call vs. something else.