Update custom-code.md
I know this is not a real edit but I didn't know how else to communicate the frustration I'm experiencing figuring out how to use this. You should write this for someone experienced with iOS or Android but who has no clue, yet, about Ionic or Capacitor. At least provide a link to more a thorough explanation. If I didn't have to do this, the lack of care of documentation would make me look elsewhere.